Set up devDependencies. Proper grunt runner.
We should never depend on package version "*"; the only wildcard should
be on minor versions, because we can be reasonably sure they won't
break APIs.

grunt-cli is not in dependencies AND it's recommended to install
globally. We don't want any external dependencies aside from node.js.
The "prestart" script was changed to run grunt without the need for
grunt-cli.